# docs/agent-tool-openapi.yml
openapi: "3.1.0"
...
info:
  title: CV SAM Service
  version: "1.0"
  description: >
    FastAPI service for image analysis, palette extraction, and SAM-based
    segmentation.  All inference endpoints accept images either as a multipart
    file upload or as a remote URL via the `image_url` query parameter.
servers:
  - url: http://localhost:5201
    description: Local / Docker Compose

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/segment:
    post:
      operationId: segmentImage
      summary: Run SAM prompted segmentation on an image
      description: >
        Uses Meta's Segment Anything Model (SAM) to generate masks for the
        region indicated by point and/or bounding-box prompts.

        **Requires the SAM model to be loaded** (`GET /api/health` → `ready: true`).
        Returns HTTP 503 while the model is still warming up.

        **Input limits:**
          - File upload: any common raster format
          - URL mode (`image_url`): http/https only, maximum **10 MB**, 10 s timeout
          - `output_format`: `masks` | `polygons` | `both`

        Provide the image via multipart `image` field **or** `image_url` query
        parameter – not both.
      parameters:
        - name: image_url
          in: query
          required: false
          schema:
            type: string
            format: uri
          description: Public http/https URL of the image (max 10 MB).
        - name: output_format
          in: query
          required: false
          schema:
            type: string
            enum: [masks, polygons, both]
            default: masks
          description: >
            `masks` – base64-encoded PNG masks only;
            `polygons` – contour point lists only;
            `both` – include both representations.
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: Input image file.
                point_coords:
                  type: string
                  description: >
                    JSON array of [x, y] point prompts, e.g. `[[320,240]]`.
                    1-indexed foreground/background determined by `point_labels`.
                point_labels:
                  type: string
                  description: >
                    JSON array of labels matching `point_coords`.
                    `1` = foreground, `0` = background.
                box:
                  type: string
                  description: >
                    JSON array `[x1, y1, x2, y2]` bounding-box prompt.
                multimask_output:
                  type: boolean
                  default: true
                  description: >
                    When `true`, SAM returns 3 candidate masks ranked by score.
      responses:
        "200":
          description: Segmentation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SegmentResponse"
              example:
                masks:
                  - "<base64-encoded PNG mask 1>"
                  - "<base64-encoded PNG mask 2>"
                  - "<base64-encoded PNG mask 3>"
                polygons: null
                scores: [0.95, 0.80, 0.70]
                processing_time_ms: 312.4
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/ModelNotReady"
  /api/segment/auto:
    post:
      operationId: segmentImageAuto
      summary: Run SAM automatic mask generation (no prompt needed)
      description: >
        Scans the entire image and returns all detectable segments, sorted by
        predicted IoU score.

        **Requires the SAM model to be loaded** (`GET /api/health` → `ready: true`).

        **Input limits:**
          - File upload: any common raster format
          - URL mode (`image_url`): http/https only, maximum **10 MB**, 10 s timeout
          - `max_masks`: 1–1000 (default 50)
      parameters:
        - name: image_url
          in: query
          required: false
          schema:
            type: string
            format: uri
          description: Public http/https URL of the image (max 10 MB).
        - name: max_masks
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
          description: Maximum number of segments returned (highest scoring first).
        - name: output_format
          in: query
          required: false
          schema:
            type: string
            enum: [masks, polygons, both]
            default: masks
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: Input image file.
      responses:
        "200":
          description: Auto-segmentation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoSegmentResponse"
              example:
                segments:
                  - mask: "<base64-encoded PNG mask>"
                    polygon: null
                    score: 0.92
                    stability_score: 0.97
                    area: 14823
                    bbox: [42.0, 18.0, 310.0, 275.0]
                count: 1
                processing_time_ms: 1840.5
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/ModelNotReady"

# ---------------------------------------------------------------------------
# Components
# ---------------------------------------------------------------------------
components:
  schemas:
...
    SegmentResponse:
      type: object
      required: [scores, processing_time_ms]
      properties:
        masks:
          type: array
          items:
            type: string
          nullable: true
          description: Base64-encoded PNG masks, one per predicted mask.
        polygons:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                type: number
          nullable: true
          description: Polygon contours for each mask as [[x,y],…].
        scores:
          type: array
          items:
            type: number
          description: Confidence score per mask (0–1).
        processing_time_ms:
          type: number
          description: Wall-clock time for prediction in milliseconds.
    AutoSegment:
      type: object
      required: [score, stability_score, area, bbox]
      properties:
        mask:
          type: string
          nullable: true
          description: Base64-encoded PNG mask.
        polygon:
          type: array
          items:
            type: array
            items:
              type: number
          nullable: true
          description: Largest contour as [[x,y],…].
        score:
          type: number
          description: Predicted IoU score (0–1).
        stability_score:
          type: number
          description: SAM stability score (0–1).
        area:
          type: integer
          description: Mask area in pixels.
        bbox:
          type: array
          items:
            type: number
          minItems: 4
          maxItems: 4
          description: "[x, y, w, h] bounding box."
    AutoSegmentResponse:
      type: object
      required: [segments, count, processing_time_ms]
      properties:
        segments:
          type: array
          items:
            $ref: "#/components/schemas/AutoSegment"
        count:
          type: integer
          description: Number of segments returned.
        processing_time_ms:
          type: number
          description: Wall-clock time for generation in milliseconds.
    ErrorDetail:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Human-readable error description.
  responses:
    ValidationError:
      description: >
        Validation error – bad parameters, unsupported URL scheme, image too
        large, fetch failure, or malformed JSON.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
          example:
            detail: "image_url scheme must be 'http' or 'https', got: 'ftp'"
    ModelNotReady:
      description: >
        SAM model has not finished loading.  Retry with exponential back-off.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
          example:
            detail: "Model not ready – please retry shortly"